# Bench parse bridge name