--- a/core/bridges/bridge_type_test.go
+++ b/core/bridges/bridge_type_test.go
@@ -1,6 +1,9 @@
 package bridges_test
 
 import (
+	"math/rand"
+	"strconv"
+	"strings"
 	"testing"
 
 	"github.com/smartcontractkit/chainlink/core/bridges"
@@ -32,6 +35,26 @@ func TestBridgeType_Authenticate(t *testing.T) {
 				assert.False(t, ok)
 			} else {
 				assert.True(t, ok)
+			}
+		})
+	}
+}
+
+func BenchmarkParseBridgeName(b *testing.B) {
+	const valid = `abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789-_`
+	for _, l := range []int{1, 10, 20, 50, 100, 1000, 10000} {
+		b.Run(strconv.Itoa(l), func(b *testing.B) {
+			sb := strings.Builder{}
+			for i := 0; i < l; i++ {
+				sb.WriteByte(valid[rand.Intn(len(valid))])
+			}
+			name := sb.String()
+			b.ResetTimer()
+			for i := 0; i < b.N; i++ {
+				_, err := bridges.ParseBridgeName(name)
+				if err != nil {
+					b.Fatalf("failed to parse %q: %v\n", name, err)
+				}
 			}
 		})
 	}