--- a/core/web/bridge_types_controller.go
+++ b/core/web/bridge_types_controller.go
@@ -31,9 +31,8 @@ func ValidateBridgeTypeNotExist(bt *bridges.BridgeTypeRequest, orm bridges.ORM) 
 	return fe.CoerceEmptyToNil()
 }
 
-// ValidateBridgeType checks that the bridge type doesn't have a duplicate
-// or invalid name or invalid url
-func ValidateBridgeType(bt *bridges.BridgeTypeRequest, orm bridges.ORM) error {
+// ValidateBridgeType checks that the bridge type has the required field with valid values.
+func ValidateBridgeType(bt *bridges.BridgeTypeRequest) error {
 	fe := models.NewJSONAPIErrors()
 	if len(bt.Name.String()) < 1 {
 		fe.Add("No name specified")
@@ -70,11 +69,11 @@ func (btc *BridgeTypesController) Create(c *gin.Context) {
 		jsonAPIError(c, http.StatusInternalServerError, err)
 		return
 	}
-	orm := btc.App.BridgeORM()
-	if e := ValidateBridgeType(btr, orm); e != nil {
+	if e := ValidateBridgeType(btr); e != nil {
 		jsonAPIError(c, http.StatusBadRequest, e)
 		return
 	}
+	orm := btc.App.BridgeORM()
 	if e := ValidateBridgeTypeNotExist(btr, orm); e != nil {
 		jsonAPIError(c, http.StatusBadRequest, e)
 		return
@@ -162,7 +161,7 @@ func (btc *BridgeTypesController) Update(c *gin.Context) {
 		jsonAPIError(c, http.StatusUnprocessableEntity, err)
 		return
 	}
-	if err := ValidateBridgeType(btr, orm); err != nil {
+	if err := ValidateBridgeType(btr); err != nil {
 		jsonAPIError(c, http.StatusBadRequest, err)
 		return
 	}