I noticed that `ParseBridgeName` was recompiling a regex on every call. Switching to a global var initialized only once is a no-brainer, but I benchmarked the difference out of curiousity too:

Before:
```
BenchmarkParseBridgeName
BenchmarkParseBridgeName/1-8 	  	  369427	      3452 ns/op
BenchmarkParseBridgeName/10-8         	  319470	      3296 ns/op
BenchmarkParseBridgeName/20-8         	  342344	      3479 ns/op
BenchmarkParseBridgeName/50-8         	  249811	      5685 ns/op
BenchmarkParseBridgeName/100-8        	  220503	      5354 ns/op
BenchmarkParseBridgeName/1000-8       	   50368	     24230 ns/op
BenchmarkParseBridgeName/10000-8      	    4558	    289070 ns/op
```

After:
```
BenchmarkParseBridgeName
BenchmarkParseBridgeName/1-8 		19158114	        75.57 ns/op
BenchmarkParseBridgeName/10-8         	 4799884	       259.1 ns/op
BenchmarkParseBridgeName/20-8         	 2807571	       443.3 ns/op
BenchmarkParseBridgeName/50-8         	 1000000	      1028 ns/op
BenchmarkParseBridgeName/100-8        	  648883	      2036 ns/op
BenchmarkParseBridgeName/1000-8       	   60464	     20087 ns/op
BenchmarkParseBridgeName/10000-8      	    4882	    281857 ns/op
```

golang.org/x/perf/cmd/benchstat
```
$ benchstat -delta-test none old.txt new.txt
name                     old time/op  new time/op  delta
ParseBridgeName/1-8      3.45µs ± 0%  0.08µs ± 0%  -97.81%
ParseBridgeName/10-8     3.30µs ± 0%  0.26µs ± 0%  -92.14%
ParseBridgeName/20-8     3.48µs ± 0%  0.44µs ± 0%  -87.26%
ParseBridgeName/50-8     5.68µs ± 0%  1.03µs ± 0%  -81.92%
ParseBridgeName/100-8    5.35µs ± 0%  2.04µs ± 0%  -61.97%
ParseBridgeName/1000-8   24.2µs ± 0%  20.1µs ± 0%  -17.10%
ParseBridgeName/10000-8   289µs ± 0%   282µs ± 0%   -2.50%
```